Reorganize pipeline IO class modules into a more consistent naming scheme.
